refactor(backend): allow aws endpoint override
To allow dev work to be self-contained, we need to remove the hard
dependency on hosted services like AWS at runtime. The first step to
this requires us to accept a parameter - an explicit RESTful endpoint -
that will be injected into Postman's AWS JavaScript clients, overriding
`AWS_REGION` in the process.

Doing so will allow developers to get Postman to connect to a specified
mock AWS service when running on their own machines, and would also
help with more invasive testing. This would also allow for external devs
to contribute, since they no longer need to have an AWS account to work
on Postman.

Note that this does not allow overrides for workers, as frameworks that
mock AWS services (eg localstack) do not have a freely available ECS API

Prep work for #284

- Add a new parameter in config - `aws.awsEndpoint`, env `AWS_ENDPOINT`
- Create a new utility function - `configureEndpoint()` - that handles
  selection of either endpoint or region for AWS config purposes
- Make use of `configureEndpoint()` in cloudwatch, S3 and SecretsManager
  - change the way we configure WinstonCloudwatch, by specifying
    `awsOptions` instead of `awsRegion`