fix(eslint): address @typescript-eslint/ban-types errors
All of these errors are the same:
error  Don't use `Function` as a type. The `Function` type accepts any function-like value.
It provides no type safety when calling the function, which can be a common source of bugs.
It also accepts things like class declarations, which will throw at runtime as they will not be called with `new`.
If you are expecting the function to accept certain arguments, you should explicitly define the function shape  @typescript-eslint/ban-types

Ref: https://github.com/typescript-cheatsheets/react/blob/main/README.md#basic-prop-types-examples