fix(app): delete render test for App.tsx
For some reason, this test on its own is failing on Travis (but succeeds
on Amplify!) and I haven't been able to find the root cause.

However, adding more tests seem to fix the failure on Travis.
Hence, I'll just batch this test up along with the other
campaign creation integration tests.