feat(frontend): improve accessibility of dashboard components (#1093)
* fix(close-button): add a title when CloseButton is used

This is easier to query by title in tests.
A neat side effect is that it adds a tooltip on hover to describe the button.

* fix(dropdown): add "listbox" role

This is semantically correct and is easier to query by role in tests.

* fix(file-input): add "button" role to the file input button

The file input button is functionally and visually a button.

Add the button role so that it's easier to query by role in tests.

* fix(dashboard/create): use TextButton for cred input buttons

Using an actual button component is semantically correct and makes it
easier to query the buttons by role in tests.

Also, the fact that TextButton has a font-size of 16px compared to the
original 14px matches the Figma design.

* fix(send-rate): add button role to header and labels to input

Add the following so that it's easier to query by role in tests:
* Button role to custom send rate header
* Labels for custom send rate input box

* fix(checkbox): add "checkbox" role

This makes it easier to query by role in tests.

* fix(dashboard): add labels to the input boxes

This makes the input boxes easier to query by role and is generally a
good idea for accessibility.

* fix(error-block): add "alert" role

This is semantically correct and makes it easier to query by role in tests.

* fix(dropdown): add "option" role to dropdown options

This makes it easier to query the options by role.

* fix(dropdown): add aria-labels to the dropdowns

THis makes the dropdowns easier to query by role in tests.

* fix(close-button): convert to button element

This makes it easier to query by role in tests.

* fix(dropdown): add aria-selected prop to selected options

This fixes the following warning:

Line 75:13:  Elements with the ARIA role "option" must have the following attributes defined: aria-selected  jsx-a11y/role-has-required-aria-props

* fix(error): remove redundant alt text for error image

Fixes the following warning when building the app:

Line 13:7:  Redundant alt attribute. Screen-readers already announce `img` tags as an image. You don’t need to use the words `image`, `photo,` or `picture` (or any specified custom words) in the alt prop  jsx-a11y/img-redundant-alt

* fix(dropdown): fix location of listbox role

The div that users actually click on is the inner div, not the outer
one.

This fixes the placement of the listbox role so that the test
environment can actually query the dropdown by role.

* fix(close-button): improve focus styles

 * Show outline only on :focus-visible, which happens when focused via a
   non-mouse/trackpad device like a keyboard
 * Change the color to light blue when focused, no matter the input
   device