chore(frontend): upgrade react-scripts to v4.0.3 (#1061)
* chore: upgrade react-scripts to v4.0.3

* chore(tsconfig): set noFallthroughCasesInSwitch to true

This is automatically set to true due to the react-scripts upgrade, with
the following stdout output:

The following changes are being made to your tsconfig.json file:
  - compilerOptions.noFallthroughCasesInSwitch to be suggested value: true (this can be changed)

This sounds like a good thing to keep, so I'll leave it as is.

Ref: https://www.typescriptlang.org/tsconfig#noFallthroughCasesInSwitch

* fix(eslint): remove @typescript-eslint/camelcase rule

This has been replaced by naming-convention [1]. Since naming-convention
does a lot more than camelcase [2], disable it entirely for now to fix the build.

Revisit naming-convention in the future to bring back the rules
camelcase provided and possibly improve upon them.

[1]: https://github.com/typescript-eslint/typescript-eslint/releases/tag/v3.0.0
[2]: https://github.com/typescript-eslint/typescript-eslint/blob/master/packages/eslint-plugin/docs/rules/naming-convention.md

* fix(eslint): disable @typescript-eslint/ban-types rule

This breaks the build currently with the following message:
Don't use `Function` as a type. The `Function` type accepts any
function-like value.

This is due to a rework of the default ban list [1].

Disable it for now and revisit later.

[1]: https://github.com/typescript-eslint/typescript-eslint/releases/tag/v3.0.0
Ref: https://github.com/typescript-eslint/typescript-eslint/blob/master/packages/eslint-plugin/docs/rules/ban-types.md

* chore(eslint): import type-only imports with `import type`

* chore(typescript-eslint): upgrade to v4.18.0

This allows our linter to match that of react-scripts.

Specifically, `npm run lint` should emit the same output as `npm run start`,
`npm start`, and `npm run build`.

For reference, react-scripts 4.0.3 uses typescript-eslint v4.5.0.

* chore(eslint): fix @typescript-eslint/no-unused-vars warnings

Most of these warnings were due to type-only imports not being stated as such.

Ref: https://github.com/typescript-eslint/typescript-eslint/blob/master/packages/eslint-plugin/docs/rules/no-unused-vars.md

* feat(eslint): ignore unused vars that are prefixed with underscores

Currently, eslint warns on unused variables even if they are prefixed
with underscores. For example, eslint warns about `_contentState` in
`VariableDecorator.tsx`.

Add a exception to the rule to address this.

Ref: https://eslint.org/docs/rules/no-unused-vars#argsignorepattern

* chore(eslint): extend react-app/jest config

This matches the default config generated by CRA.

* fix(eslint): disable @typescript-eslint/explicit-module-boundary-types

There are 138 warnings for this. Most of these warnings are regarding
the lack of return values for React components.

More research is needed to determine the right way to address this, be
it by introducing explicit return values, or adjusting the rulesets.

Until then, disable this for now.

Refs:
https://github.com/typescript-cheatsheets/react#function-components
https://github.com/typescript-eslint/typescript-eslint/blob/master/packages/eslint-plugin/docs/rules/explicit-module-boundary-types.md

* fix: explicitly disable no-console rule for intentional console.log

* fix: disable eslint warning when using useCallback with lodash debounce

The warning:
15:25  warning  React Hook useCallback received a function whose dependencies are unknown. Pass an inline function instead  react-hooks/exhaustive-deps

This isn't a concern for us since the debounced function returned by
_.debounce() will never change.

Disable the warning.

* chore(eslint): upgrade eslint to v7

This matches the version used internally by react-scripts v4.

* fix(eslint): disable no-undef rule for .eslintrc.js

ESLint was complaining that this module should use ES6 modules, but
that's impossible.

This must be a commonjs module since it's not transpiled as part of
the build.

* Revert "fix(eslint): disable @typescript-eslint/ban-types rule"

0 warnings left -- time to revisit this.

This reverts commit b92392b618df68cf158abe0f85c7936769cbd027.

* fix(eslint): address @typescript-eslint/ban-types errors

All of these errors are the same:
error  Don't use `Function` as a type. The `Function` type accepts any function-like value.
It provides no type safety when calling the function, which can be a common source of bugs.
It also accepts things like class declarations, which will throw at runtime as they will not be called with `new`.
If you are expecting the function to accept certain arguments, you should explicitly define the function shape  @typescript-eslint/ban-types

Ref: https://github.com/typescript-cheatsheets/react/blob/main/README.md#basic-prop-types-examples

* chore(eslint): remove redundant @typescript-eslint/eslint-recommended

@typescript-eslint/recommended already inherits this ruleset.

Ref: https://github.com/typescript-eslint/typescript-eslint/releases/tag/v3.0.0

* chore(file-input): refactor onFileSelected to accept a FileList param

This is cleaner than weirdly casting a `FileList | null` type into a
`File[]` type.

* chore(frontend): upgrade y18n from 4.0.0 to 4.0.1

Co-authored-by: Lam Kee Wei <keewei.lam@gmail.com>