fix(dashboard): only call setIsUploading if component is mounted
When running tests, React logs the following warnings:

Warning: Can't perform a React state update on an unmounted component. This is a no-op, but it indicates a memory leak in your application. To fix, cancel all subscriptions and asynchronous tasks in a useEffect cleanup function.
        in TelegramRecipients (at TelegramCreate.tsx:39)

Warning: Can't perform a React state update on an unmounted component. This is a no-op, but it indicates a memory leak in your application. To fix, cancel all subscriptions and asynchronous tasks in a useEffect cleanup function.
        in SMSRecipients (at SMSCreate.tsx:41)

This is because setIsUploading is being called even after the component
is unmounted. Move the setIsUploading call out of the finally block so
that the return statement in the try block holds.

Note: this mirrors what is being done in EmailRecipients.tsx as well.
A better long term solution would probably be to deduplicate the code.