fix(campaigns): only setState if component is mounted
When running tests, React logs the following errors:

Warning: Can't perform a React state update on an unmounted component. This is a no-op, but it indicates a memory leak in your application. To fix, cancel all subscriptions and asynchronous tasks in a useEffect cleanup function.
        in Campaigns (created by Context.Consumer)

Fix this by checking if the component is still mounted before calling
setState.