fix(util): Correct type signature of `replaceNewLines`
The input for `replaceNewLines` can actually be undefined, so make it
explicit.

Furthermore, use the nullish coalescing operator (??) instead of the
logical OR operator (||) to determine if the input is undefined. In
practice, this doesn't make a difference for us, but it is good to be
clear about our intentions here.

Ref on the nullish coalescing operator: https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Operators/Nullish_coalescing_operator