feat: only prepend 'Copy of' if the campaign name doesn't have it
If a user duplicates a campaign, then duplicates the duplicate, and does
it over and over again, the resulting campaign name becomes something
like "Copy of Copy of Copy of Copy of ..."

This is pretty ridiculous, and doesn't really make sense to have.
All these campaigns are really just similar copies of the original campaign anyway.

Hence, only prepend "Copy of " if the campaign name doesn't already have
it.