Revert "middleware: Don't expose the custom address to the user during fallback"
A side effect of this is that the email section in the settings tab will
be reverted to the request form (like when they don't have a custom from
registered).

This might cause users to be alarmed that their custom from is now
missing.

Let's just rely on the error message from the subsequent PUT request.

This reverts commit 40228df7ac23c72b5a148e5edf6c046da9be0efd.