fix(frontend): fix frontend test flakiness (#1162)
* fix(frontend): increase jest timeout to 20s

We were experiencing flaky tests as described in issue #1161.

A timeout of 10s works fine on a local MBP 16". However, given that
the CI servers are likely much weaker, 2x the timeout to be on the safe
side.

* feat: configure the timeout using an env var

This is so that we can configure the timeout based on the test
environment.

* fix(test): add /settings/announcement-version API mock

* fix: add a dummy API mock for /settings/announcement-version

* fix: disable announcements for all tests

We don't have any tests that test the announcement modal.
Hence, disable it for now.

A longer-term fix would be to add announcement_version to the initial
state of the API mock.

* fix: correct the type of isActive

* chore: don't run tests on Amplify

We only rely on Amplify for deployments, so there isn't a major reason
to run tests there.

The main reason is to block the deployment if a test fails, but any test
failures should've been caught on Travis during development.

If a failure does happen during deployment, we can manually redeploy an
older release.