Fix tough-cookie Prototype Pollution vulnerability
project used the package tough-cookie before 4.1.3 are vulnerable to Prototype Pollution due to improper handling of Cookies when using CookieJar in `rejectPublicSuffixes=false` mode. This issue arises from the manner in which the objects are initialized.

**PoC:**
```js
// PoC.js async function main(){ var tough = require("tough-cookie"); var cookiejar = new tough.CookieJar(undefined,{rejectPublicSuffixes:false}); // Exploit cookie await cookiejar.setCookie( "Slonser=polluted; Domain=__proto__; Path=/notauth", "https://__proto__/admin" ); // normal cookie var cookie = await cookiejar.setCookie( "Auth=Lol; Domain=google.com; Path=/notauth", "https://google.com/" ); //Exploit cookie var a = {}; console.log(a["/notauth"]["Slonser"]) } main();
```
**Details:**
Prototype Pollution is a vulnerability affecting JavaScript. Prototype Pollution refers to the ability to inject properties into existing JavaScript language construct prototypes, such as objects. JavaScript allows all Object attributes to be altered, including their magical attributes such as __proto__, constructor and prototype. An attacker manipulates these attributes to overwrite, or pollute, a JavaScript application object prototype of the base object by injecting other values. Properties on the Object.prototype are then inherited by all the JavaScript objects through the prototype chain. When that happens, this leads to either denial of service by triggering JavaScript exceptions, or it tampers with the application source code to force the code path that the attacker injects, thereby leading to remote code execution.

There are two main ways in which the pollution of prototypes occurs:
 * Unsafe Object recursive merge
 * Property definition by path

**Impact:**
CVE-2023-26136
CWE-1321
**`CVSS:3.1/AV:N/AC:L/PR:N/UI:N/S:U/C:L/I:L/A:N`**

## Problem

_What problem are you trying to solve? What issue does this close?_

Closes [insert issue #]

## Solution

_How did you solve the problem?_

## Deployment Checklist

_Any tasks that need to be done on the environment before releasing of this PR (e.g. Setting environment variables, running migrations, etc). If there's no task to be done, put "N/A"_

- [x] First task
  - [ ] Sub-task 1
  - [ ] Sub-task 2
- [ ] Second task